Adopt new naming convention for Geant4 assembly imprints to allow recuperation of mother/daughter copy numbers
