+++ DDG4/include/DDG4/Geant4AssemblyVolume.h
@@ -1,6 +1,3 @@
-#ifndef DDG4_GEANT4ASSEMBLYVOLUME_H
-#define DDG4_GEANT4ASSEMBLYVOLUME_H
-
 //==========================================================================
 //  AIDA Detector description implementation 
 //--------------------------------------------------------------------------
@@ -13,6 +10,8 @@
 // Author     : M.Frank
 //
 //==========================================================================
+#ifndef DDG4_GEANT4ASSEMBLYVOLUME_H
+#define DDG4_GEANT4ASSEMBLYVOLUME_H
 
 // Disable diagnostics for ROOT dictionaries
 #ifdef __clang__
@@ -28,12 +27,21 @@
 #pragma clang diagnostic pop
 #endif
 
+// ROOT includes
+#include "TGeoNode.h"
+
+/// C/C++ include files
+#include <vector>
+
 /// Namespace for the AIDA detector description toolkit
 namespace dd4hep {
 
   /// Namespace for the Geant4 based simulation part of the AIDA detector description toolkit
   namespace sim {
 
+    /// Forward declarations
+    class Geant4GeometryInfo;
+    
     /// Hack! Wrapper around G4AssemblyVolume to access protected members.
     /**
      *  \author  M.Frank
